Further refactoring of (SafeNet)ServeContent
